requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
psutil==5.8.0
pyautogui==0.9.52
python-telegram-bot==13.4.1
pywin32==300
requests==2.25.1
urllib3==1.26.4